chore: metaデータをlayoutに移動
